Fixed some things

+++ src/events/messageCreate.ts
@@ -2,14 +2,18 @@
 
 import { Event } from "fero-dc";
 import { User } from "../models/User";
+import { Guild, IGuild } from "../models/Guild";
 import { getBannedWord } from "../scripts/getBannedWord";
-
 export default {
   event: "messageCreate",
   run: async (client, message) => {
-    if (message.author.bot) return;
+    if (message.author.bot || !message.guild) return;
 
     if (getBannedWord(message) && message.guild) return message.delete();
+
+    const guildModel: IGuild = await Guild.findOne({ _id: message.guild.id });
+
+    if (!(guildModel?.features?.scales ?? true)) return;
 
     try {
       const randomAmount = Math.floor(Math.random() * 50) + 1;
